kpmg_assignment/part_2/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gradio==4.8.0
langchain==0.3.0
langchain-openai==0.2.2
langgraph==0.5.3
pydantic==2.7.4
python-dotenv==1.0.0
faiss-cpu==1.7.4
beautifulsoup4==4.12.2
lxml==4.9.3
httpx==0.25.2
python-multipart==0.0.6